requirements.txt
numpy<2
torch>=2.0.0
einops==0.8.1
huggingface_hub==0.33.1
matplotlib==3.9.3
pandas==2.2.2
tqdm==4.67.1
safetensors==0.6.2
python-binance==1.0.19
python-dotenv==1.0.1
requests==2.31.0
websockets==12.0
psutil
nvidia-ml-py
feedparser>=6.0.0
beautifulsoup4>=4.10.0
